+++ internal/assets/embed.go
@@ -1,0 +1,6 @@
+package assets
+
+import "embed"
+
+//go:embed views/* assets/*
+var EmbeddedFiles embed.FS
