+++ internal/k8s/client.go
@@ -1,0 +1,37 @@
+package k8s
+
+import (
+	"fmt"
+
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
+)
+
+// GetKubeClient creates a Kubernetes client from either a kubeconfig file or in-cluster configuration.
+func GetKubeClient(kubeconfigPath string) (kubernetes.Interface, error) {
+	var config *rest.Config
+	var err error
+
+	if kubeconfigPath != "" {
+		// Use the provided kubeconfig file
+		config, err = clientcmd.BuildConfigFromFlags("", kubeconfigPath)
+		if err != nil {
+			return nil, fmt.Errorf("unable to load kubeconfig from %s: %w", kubeconfigPath, err)
+		}
+	} else {
+		// Use in-cluster configuration
+		config, err = rest.InClusterConfig()
+		if err != nil {
+			return nil, fmt.Errorf("unable to load in-cluster config: %w", err)
+		}
+	}
+
+	// Create the clientset
+	clientset, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		return nil, fmt.Errorf("unable to create a Kubernetes client: %w", err)
+	}
+
+	return clientset, nil
+}
